refactor: split commands.rs into a commands/ module tree
commands.rs had grown to 2635 lines. Split it into one submodule per
command domain with no behavior change — pure code movement plus the
per-file import headers and module wiring.

  commands/
    mod.rs                     re-exports the public handler surface
    util.rs                    parse_time_string, run_watch_loop, is_auth_error
    order/trade/leverage/margin/account/config/auth/market/stream/block.rs
    dashboard.rs  portfolio.rs
    maker/  mod.rs   handle_maker, run_maker, error/adoption helpers
            feed.rs  FeedState, spawn_market_feed, market_snapshot
            cycle.rs maker_cycle, cancel_all_with_retry
            output.rs emit_maker_cycle, log_maker_event

main.rs and the tests are untouched: `pub mod commands;` stays, and the
handlers keep their `commands::handle_*` paths via re-exports. Cross-
module helpers use `pub(super)`; maker submodules reach shared parent
items through `super::`. Unit tests moved next to the code they cover
(parse-time → util, rejection/adoption → maker).

Verified: `cargo test` (133, unchanged), `cargo clippy --workspace
--all-targets -- -D warnings`, `cargo fmt --check` all green; maker
paper-mode smoke unchanged.
